Update URLs and names to reflect changed name.
According to #42.
This is a follow-up to #61 and updates remaining URLs. Some of the currently listed URLs are broken (404), and some redirect. All new URLs are working and do not redirect.
It seems that commit 2379ba34a6d5578befc5f9d1e8ea532d281847be accidentally reverted some changes of #61.